Refactor alert dialog imports
